requirements.txt
Bio>=1.8.0
biopython>=1.81
h5py>=3.7.0
matplotlib>=3.7.1
numpy>=1.24.3
pandas>=1.5.3
scikit_learn>=1.2.2
scipy>=1.16.1
seaborn>=0.13.2
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.65.2
umap_learn>=0.5.3
zarr>=2.18.3